requirements.txt
fastapi==0.95.0
uvicorn==0.23.0
sqlalchemy==2.0.10
pydantic==1.10.2
passlib[bcrypt]==1.7.4
python-dotenv==0.21.0
alembic==1.10.2
psycopg2==2.9.3
Package           Version
----------------- --------
aiosmtplib        3.0.2
alembic           1.15.2
annotated-types   0.7.0
anyio             4.9.0
APScheduler       3.11.0
bcrypt            4.3.0
blinker           1.9.0
cffi              1.17.1
click             8.1.8
colorama          0.4.6
cryptography      44.0.2
dnspython         2.7.0
ecdsa             0.19.1
email_validator   2.2.0
fastapi           0.115.12
fastapi-mail      1.4.2
greenlet          3.2.1
h11               0.16.0
idna              3.10
Jinja2            3.1.6
Mako              1.3.10
MarkupSafe        3.0.2
passlib           1.7.4
pip               25.1
psycopg2-binary   2.9.10
pyasn1            0.4.8
pycparser         2.22
pydantic          2.11.4
pydantic_core     2.33.2
pydantic-settings 2.9.1
python-dotenv     1.1.0
python-jose       3.4.0
python-multipart  0.0.20
rsa               4.9.1
six               1.17.0
sniffio           1.3.1
SQLAlchemy        2.0.40
starlette         0.46.2
typing_extensions 4.13.2
typing-inspection 0.4.0
tzdata            2025.2
tzlocal           5.3.1
uvicorn           0.34.2
